+++ CHANGELOG.md
@@ -123,6 +123,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated specification `engine-core` (engine)
 - Updated implementation plan (engine)
 - Updated a specification (engine)
+- Updated 8 specifications (engine)
 
 ### Added